Created a tests folder + moved existing tests in it + modified tests script
